refactor: Update Bio component to use localized data
This commit updates the Bio component to use localized data from the `bio-info.ts` file. Instead of hardcoding the name, profession, and other details, the component now imports the data from the `bioInfo` object. This change improves maintainability and allows for easy localization of the Bio section in different languages.