refactor: Add BlogLinksList component and update page layout
This commit adds the BlogLinksList component and updates the page layout in the blog/[slug]/page.tsx file. The BlogLinksList component is responsible for rendering a list of blog links with their respective technologies. The page layout is modified to include the BlogLinksList component, displaying the technologies and links associated with the blog post. This enhancement improves the functionality and user experience of the blog page.